suse-commit: a59f43783aed617963fd09c6e82731bedc6c2308

+++ drivers/acpi/thermal.c
@@ -84,11 +84,11 @@ static int acpi_thermal_remove (struct acpi_device *device, int type);
 static int acpi_thermal_state_open_fs(struct inode *inode, struct file *file);
 static int acpi_thermal_temp_open_fs(struct inode *inode, struct file *file);
 static int acpi_thermal_trip_open_fs(struct inode *inode, struct file *file);
-static int acpi_thermal_write_trip_points (struct file*,const char *,size_t,loff_t *);
+static ssize_t acpi_thermal_write_trip_points (struct file*,const char *,size_t,loff_t *);
 static int acpi_thermal_cooling_open_fs(struct inode *inode, struct file *file);
-static int acpi_thermal_write_cooling_mode (struct file*,const char *,size_t,loff_t *);
+static ssize_t acpi_thermal_write_cooling_mode (struct file*,const char *,size_t,loff_t *);
 static int acpi_thermal_polling_open_fs(struct inode *inode, struct file *file);
-static int acpi_thermal_write_polling(struct file*,const char *,size_t,loff_t *);
+static ssize_t acpi_thermal_write_polling(struct file*,const char *,size_t,loff_t *);
 
 static struct acpi_driver acpi_thermal_driver = {
 	.name =		ACPI_THERMAL_DRIVER_NAME,
@@ -881,7 +881,7 @@ static int acpi_thermal_trip_open_fs(struct inode *inode, struct file *file)
 	return single_open(file, acpi_thermal_trip_seq_show, PDE(inode)->data);
 }
 
-static int
+static ssize_t
 acpi_thermal_write_trip_points (
         struct file		*file,
         const char		*buffer,
@@ -950,7 +950,7 @@ static int acpi_thermal_cooling_open_fs(struct inode *inode, struct file *file)
 							PDE(inode)->data);
 }
 
-static int
+static ssize_t
 acpi_thermal_write_cooling_mode (
 	struct file		*file,
 	const char		*buffer,
@@ -1011,7 +1011,7 @@ static int acpi_thermal_polling_open_fs(struct inode *inode, struct file *file)
 							PDE(inode)->data);
 }
 
-static int
+static ssize_t
 acpi_thermal_write_polling (
 	struct file		*file,
 	const char		*buffer,
